BEH: add :no-inherited-members: to automodule
We use :inherited-members: in source/conf.py:autodoc_default_options in
projects to document inherited members in classes. The docs [1] suggest
that this setting is applied only to autoclass directives. However, it
seems to get applied to all autoXXX directives, not just autoclass. This
causes all module content (classes, functions, ...) to end up on a single
page. Using :no-inherited-members" in all automodule directives prevents
that.

[1] https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#directive-autoclass